requirements.txt
alembic
pydantic
PyMySQL
python-dotenv
python-telegram-bot
sqlalchemy
